requirements.txt
Appium-Python-Client~=4.5.0
pytest-bdd~=8.1.0
selenium~=4.28.0
pytest~=8.3.4
pytest_metadata
pytest-html
opencv-python
scikit-image